+++ Sharp.BlobStorage/Internal/UriExtensions.cs
@@ -1,4 +1,20 @@
-﻿using System;
+/*
+    Copyright 2020 Jeffrey Sharp
+
+    Permission to use, copy, modify, and distribute this software for any
+    purpose with or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+    MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+
+using System;
 
 namespace Sharp.BlobStorage.Internal
 {
@@ -31,6 +47,52 @@ namespace Sharp.BlobStorage.Internal
             var builder = new UriBuilder(uri);
             builder.Path += "/";
             return builder.Uri;
+        }
+
+        /// <summary>
+        ///   Converts the current absolute URI to a relative URI by removing
+        ///   the specified base URI.
+        /// </summary>
+        /// <param name="uri">
+        ///   The absolute URI in which to remove the base URI.
+        /// </param>
+        /// <param name="baseUri">
+        ///   The base URI to remove from <paramref name="uri"/>.
+        /// </param>
+        /// <returns>
+        ///   A copy of <paramref name="uri"/> from which
+        ///   <paramref name="baseUri"/> has been removed.
+        /// </returns>
+        /// <exception cref="ArgumentNullException">
+        ///   One or more of
+        ///     <paramref name="uri"/> or
+        ///     <paramref name="baseUri"/>
+        ///     is <c>null</c>.
+        /// </exception>
+        /// <exception cref="ArgumentException">
+        ///   One or more of
+        ///     <paramref name="uri"/> or
+        ///     <paramref name="baseUri"/>
+        ///     is a relative URI; or,
+        ///   <paramref name="baseUri"/> is not a base of
+        ///     <paramref name="uri"/>.
+        /// </exception>
+        public static Uri ToRelative(this Uri uri, Uri baseUri)
+        {
+            if (uri == null)
+                throw new ArgumentNullException(nameof(uri));
+            if (baseUri == null)
+                throw new ArgumentNullException(nameof(baseUri));
+
+            if (!uri.IsAbsoluteUri)
+                throw UriNotAbsoluteError(uri, nameof(uri));
+            if (!baseUri.IsAbsoluteUri)
+                throw UriNotAbsoluteError(baseUri, nameof(baseUri));
+
+            if (!baseUri.IsBaseOf(uri))
+                throw UriNotRequiredBaseError(uri, baseUri, nameof(uri));
+
+            return baseUri.MakeRelativeUri(uri);
         }
 
         /// <summary>