+++ CHANGES.md
@@ -8,7 +8,9 @@ Most lines should begin with one of these words:
 (none)
 
 ## [1.0.0](https://github.com/sharpjs/Sharp.BlobStorage/compare/v0.1.0..v1.0.0)
-Future release.
+- Change underlying SDK
+  from [WindowsAzure.Storage](https://www.nuget.org/packages/WindowsAzure.Storage/)
+  to   [Azure.Storage.Blobs](https://www.nuget.org/packages/Azure.Storage.Blobs) 12.6.0.
 
 ## [0.1.0](https://github.com/sharpjs/Sharp.BlobStorage/tree/v0.1.0)
 Initial release.